--- src/main/kotlin/team/react/school_chat/auth/domain/Friend.kt
@@ -1,20 +1,0 @@
-package team.react.school_chat.auth.domain
-
-import jakarta.persistence.*
-
-@Entity
-@Table(name = "friend")
-data class Friend(
-    @EmbeddedId
-    val id: FriendId,
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @MapsId("ownerEmail")
-    @JoinColumn(name = "owner_email", nullable = false)
-    var owner: Member,
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @MapsId("friendEmail")
-    @JoinColumn(name = "friend_email", nullable = false)
-    var friend: Member,
-)
