--- src/main/kotlin/team/react/school_chat/auth/dto/MemberUpdateRequest.kt
@@ -1,8 +1,0 @@
-package team.react.school_chat.auth.dto
-
-import team.react.school_chat.auth.domain.Member
-
-data class MemberUpdateRequest(
-    val name: String,
-    val picture: String
-)
