--- src/main/kotlin/team/react/school_chat/chat/Chat.kt
@@ -1,11 +1,0 @@
-package team.react.school_chat.chat
-
-import java.time.LocalDateTime
-
-data class Chat(
-        val id: String? = null,
-        val userName: String? = null,
-        val content: String? = null,
-        var createTime: LocalDateTime? = LocalDateTime.now(),
-        val roomId: String? = null
-)
