--- src/main/kotlin/team/react/school_chat/chat/config/CorsConfig.kt
@@ -1,16 +1,0 @@
-package team.react.school_chat.chat.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class CorsConfig :WebMvcConfigurer{
-    override fun  addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-                .allowedOrigins("http://127.0.0.1:5500")
-                .allowedMethods("GET", "POST", "PUT", "DELETE")
-                .allowedHeaders("*")
-                .allowCredentials(true)
-    }
-}
