--- src/main/kotlin/team/react/school_chat/chat/config/WebSocketConfig.kt
@@ -1,26 +1,0 @@
-package team.react.school_chat.chat.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry
-import org.springframework.scheduling.annotation.EnableScheduling
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-import org.springframework.web.socket.WebSocketHandler
-import org.springframework.web.socket.config.annotation.*
-
-@Configuration
-@EnableWebSocket
-@EnableWebSocketMessageBroker
-public class WebSocketConfig (private val webSocketHandler: WebSocketHandler):WebSocketConfigurer, WebSocketMessageBrokerConfigurer {
-    override fun registerWebSocketHandlers(registry: WebSocketHandlerRegistry) {
-        registry.addHandler(webSocketHandler, "/ws/chat").setAllowedOrigins("http://127.0.0.1:5500")
-    }
-
-    override fun registerStompEndpoints(registry: StompEndpointRegistry) {
-        registry.addEndpoint("/chat").setAllowedOrigins("http://127.0.0.1:5500").withSockJS()
-    }
-    override fun configureMessageBroker(config: MessageBrokerRegistry) {
-        config.setApplicationDestinationPrefixes("/app")
-        config.enableSimpleBroker("/topic")
-    }
-}
