--- src/main/kotlin/team/react/school_chat/chatroom/ChatRoom.kt
@@ -1,8 +1,0 @@
-package team.react.school_chat.chatroom
-
-data class ChatRoom (
-        val id : String? = null,
-        val name : String? = null,
-        val email: String? = null,
-        var age: Int? = null
-)
