--- src/main/kotlin/team/react/school_chat/chatroom/controller/ChatRoomController.kt
@@ -1,12 +1,0 @@
-package team.react.school_chat.chatroom.controller
-
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class ChatRoomController {
-    @PostMapping()
-    fun createPersonalChatRoom(){
-
-    }
-}
