+++ src/main/kotlin/team/react/school_chat/domain/auth/dto/MemberUpdateRequest.kt
@@ -1,0 +1,8 @@
+package team.react.school_chat.domain.auth.dto
+
+import team.react.school_chat.domain.auth.domain.Member
+
+data class MemberUpdateRequest(
+    val name: String,
+    val picture: String
+)
