+++ src/main/kotlin/team/react/school_chat/domain/chat/controller/ChatRoomController.kt
@@ -1,0 +1,12 @@
+package team.react.school_chat.domain.chat.controller
+
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class ChatRoomController {
+    @PostMapping()
+    fun createPersonalChatRoom(){
+
+    }
+}
