+++ src/main/kotlin/team/react/school_chat/global/utils/JwtAuthenticationFilter.kt
@@ -1,0 +1,59 @@
+package team.react.school_chat.global.utils
+
+import jakarta.servlet.FilterChain
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.springframework.http.HttpStatus
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
+import org.springframework.security.core.authority.SimpleGrantedAuthority
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.stereotype.Component
+import org.springframework.util.AntPathMatcher
+import org.springframework.web.filter.OncePerRequestFilter
+
+@Component
+class JwtAuthenticationFilter(
+    private val jwtProvider: JwtProvider,
+) : OncePerRequestFilter() {
+    private val protectedPaths = listOf("/api/user/email", "/member/**")
+
+    private val pathMatcher = AntPathMatcher()
+
+    override fun shouldNotFilter(request: HttpServletRequest): Boolean {
+        return !protectedPaths.any { pathMatcher.match(it, request.requestURI) }
+    }
+
+    override fun doFilterInternal(
+        request: HttpServletRequest,
+        response: HttpServletResponse,
+        filterChain: FilterChain
+    ) {
+        val token = resolveToken(request)
+
+        if (token != null && jwtProvider.validateToken(token)) {
+            val email = jwtProvider.getEmailFromToken(token)
+            val authentication = UsernamePasswordAuthenticationToken(email, null, listOf(SimpleGrantedAuthority("ROLE_USER")))
+            SecurityContextHolder.getContext().authentication = authentication
+            filterChain.doFilter(request, response)
+            return
+        }
+
+        response.sendError(HttpStatus.UNAUTHORIZED.value(), "Invalid or Expired Token")
+
+
+
+    }
+
+    private fun resolveToken(request: HttpServletRequest): String? {
+        val bearerToken = request.getHeader("Authorization")
+        return if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
+            bearerToken.substring(7)
+        } else null
+    }
+
+    private fun handleException(response: HttpServletResponse, message: String, status: HttpStatus) {
+        response.status = status.value()
+        response.contentType = "application/json"
+        response.writer.write("{\"error\": \"$message\"}")
+    }
+}
