+++ src/main/kotlin/team/react/school_chat/global/utils/JwtProvider.kt
@@ -1,0 +1,41 @@
+package team.react.school_chat.global.utils
+
+import io.jsonwebtoken.Jwts
+import io.jsonwebtoken.SignatureAlgorithm
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.stereotype.Component
+import io.jsonwebtoken.security.Keys
+import java.util.Date
+
+
+@Component
+class JwtProvider(
+    @Value("\${jwt.secret}")
+    private val secretKey: String,
+) {
+    private val key = Keys.hmacShaKeyFor(secretKey.toByteArray())
+
+    fun generateToken(email: String): String {
+        return Jwts.builder()
+            .setSubject(email)
+            .setIssuedAt(Date())
+            .setExpiration(Date(System.currentTimeMillis() + 1000 * 60 * 60)) // 1시간 유효
+            .signWith(key, SignatureAlgorithm.HS256)
+            .compact()
+    }
+
+    fun validateToken(token: String): Boolean {
+        return try {
+            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token)
+            true
+        } catch (e: Exception) {
+            false
+        }
+    }
+
+    fun getEmailFromToken(token: String): String {
+        return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).body.subject
+    }
+
+
+}
